Refactor out more complicated attribute archetype logic in preparation for yang autogen replacement
